Consolidate time limit definitions in the Regulations. Closes #611.
This moves many details from the Guidelines to the Regulations. It also includes
two functional changes:

- Time limits must be announced before the competition, rather than just before the
  round.
- Only one cumulative time limit can apply to to a given round.

Both of these are almost always (or always?) followed in practice.

Note that this change "recycles" A1a2++) for a Guideline that isn't particular related to the old A1a2++. We don't usually do this, but it avoids creating a Guideline with the loooooong label of A2a++++++ and I don't foresee it causing issues here.